Epub cleanup.

+++ src/epub.hpp
@@ -21,7 +21,7 @@ public:
 	bool load(const std::string& fname);
 	int get_num_sections() const;
 	std::string get_section_text(epub_section& section);
-	epub_section parse_section(unsigned int n, std::vector<std::string>* lines);
+	epub_section parse_section(size_t n, std::vector<std::string>* lines);
 	const std::string& title() const{return title_;}
 	const std::string& author() const{return author_;}
